dividing code in chat container with code related to messages and waku chat with code related to component handle like open close settings etc
